+++ src/main/java/io/github/astrapi69/gen/tree/handler/ITreeNodeHandlerExtensions.java
@@ -24,15 +24,14 @@
  */
 package io.github.astrapi69.gen.tree.handler;
 
-import io.github.astrapi69.design.pattern.visitor.Visitor;
-import io.github.astrapi69.gen.tree.api.ITreeNode;
-import lombok.NonNull;
-
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.LinkedHashSet;
 import java.util.List;
 import java.util.concurrent.atomic.AtomicReference;
+
+import io.github.astrapi69.gen.tree.api.ITreeNode;
+import lombok.NonNull;
 
 /**
  * The class {@link ITreeNodeHandlerExtensions} provides handler methods for the class
@@ -47,7 +46,7 @@ public class ITreeNodeHandlerExtensions
 	 * @param <V>
 	 *            the generic type of the value
 	 * @param <T>
-	 *            the generic type of the id of the node
+	 *            the generic type of the concrete tree node
 	 * @param treeNode
 	 *            the tree node
 	 * @return Returns all siblings of the given {@link ITreeNode} object
@@ -71,7 +70,7 @@ public class ITreeNodeHandlerExtensions
 	 * @param <V>
 	 *            the generic type of the value
 	 * @param <T>
-	 *            the generic type of the id of the node
+	 *            the generic type of the concrete tree node
 	 * @param treeNode
 	 *            the tree node
 	 * @return the root from the given {@link ITreeNode} object
@@ -98,7 +97,7 @@ public class ITreeNodeHandlerExtensions
 	 * @param <V>
 	 *            the generic type of the value
 	 * @param <T>
-	 *            the generic type of the id of the node
+	 *            the generic type of the concrete tree node
 	 * @param currentTreeNode
 	 *            the tree node
 	 * @return the next sibling of the given {@link ITreeNode} object or null if the given
@@ -135,7 +134,7 @@ public class ITreeNodeHandlerExtensions
 	 * @param <V>
 	 *            the generic type of the value
 	 * @param <T>
-	 *            the generic type of the id of the node
+	 *            the generic type of the concrete tree node
 	 *
 	 * @param currentTreeNode
 	 *            the tree node
@@ -168,7 +167,7 @@ public class ITreeNodeHandlerExtensions
 	 * @param <V>
 	 *            the generic type of the value
 	 * @param <T>
-	 *            the generic type of the id of the node
+	 *            the generic type of the concrete tree node
 	 *
 	 * @param treeNode
 	 *            the tree node
@@ -191,7 +190,7 @@ public class ITreeNodeHandlerExtensions
 	 * @param <V>
 	 *            the generic type of the value
 	 * @param <T>
-	 *            the generic type of the id of the node
+	 *            the generic type of the concrete tree node
 	 * @param parentTreeNode
 	 *            the parent tree node
 	 * @param child
@@ -218,7 +217,7 @@ public class ITreeNodeHandlerExtensions
 	 * @param <V>
 	 *            the generic type of the value
 	 * @param <T>
-	 *            the generic type of the id of the node
+	 *            the generic type of the concrete tree node
 	 * @param parentTreeNode
 	 *            the parent tree node
 	 * @param child
@@ -239,7 +238,7 @@ public class ITreeNodeHandlerExtensions
 	 * @param <V>
 	 *            the generic type of the value
 	 * @param <T>
-	 *            the generic type of the id of the node
+	 *            the generic type of the concrete tree node
 	 * @param parentTreeNode
 	 *            the parent tree node
 	 * @param child
@@ -259,7 +258,7 @@ public class ITreeNodeHandlerExtensions
 	 * @param <V>
 	 *            the generic type of the value
 	 * @param <T>
-	 *            the generic type of the id of the node
+	 *            the generic type of the concrete tree node
 	 * @param parentTreeNode
 	 *            the parent tree node
 	 * @param children
@@ -277,7 +276,7 @@ public class ITreeNodeHandlerExtensions
 	 * @param <V>
 	 *            the generic type of the value
 	 * @param <T>
-	 *            the generic type of the id of the node
+	 *            the generic type of the concrete tree node
 	 * @param parentTreeNode
 	 *            the parent tree node
 	 */
@@ -294,7 +293,7 @@ public class ITreeNodeHandlerExtensions
 	 * @param <V>
 	 *            the generic type of the value
 	 * @param <T>
-	 *            the generic type of the id of the node
+	 *            the generic type of the concrete tree node
 	 * @param treeNode
 	 *            the tree node
 	 * @return true, if the given {@link ITreeNode} is the root {@link ITreeNode} object
@@ -305,12 +304,29 @@ public class ITreeNodeHandlerExtensions
 	}
 
 	/**
+	 * Checks if the given {@link ITreeNode} object has a next sibling {@link ITreeNode} object
+	 *
+	 * @param <V>
+	 *            the generic type of the value
+	 * @param <T>
+	 *            the generic type of the concrete tree node
+	 * @param treeNode
+	 *            the tree node
+	 * @return true, if the given {@link ITreeNode} object has a next sibling {@link ITreeNode}
+	 *         object otherwise false
+	 */
+	public static <V, T extends ITreeNode<V, T>> boolean hasNextSibling(final @NonNull T treeNode)
+	{
+		return treeNode.getNextSibling() != null;
+	}
+
+	/**
 	 * Checks if the given {@link ITreeNode} object has a parent {@link ITreeNode} object
 	 *
 	 * @param <V>
 	 *            the generic type of the value
 	 * @param <T>
-	 *            the generic type of the id of the node
+	 *            the generic type of the concrete tree node
 	 * @param treeNode
 	 *            the tree node
 	 * @return true, if the given {@link ITreeNode} object has a parent {@link ITreeNode} object
@@ -322,12 +338,30 @@ public class ITreeNodeHandlerExtensions
 	}
 
 	/**
+	 * Checks if the given {@link ITreeNode} object has a previous sibling {@link ITreeNode} object
+	 *
+	 * @param <V>
+	 *            the generic type of the value
+	 * @param <T>
+	 *            the generic type of the concrete tree node
+	 * @param treeNode
+	 *            the tree node
+	 * @return true, if the given {@link ITreeNode} object has a previous sibling {@link ITreeNode}
+	 *         object otherwise false
+	 */
+	public static <V, T extends ITreeNode<V, T>> boolean hasPreviousSibling(
+		final @NonNull T treeNode)
+	{
+		return treeNode.getPreviousSibling() != null;
+	}
+
+	/**
 	 * Checks if the given {@link ITreeNode} object is a node
 	 *
 	 * @param <V>
 	 *            the generic type of the value
 	 * @param <T>
-	 *            the generic type of the id of the node
+	 *            the generic type of the concrete tree node
 	 * @param treeNode
 	 *            the tree node
 	 * @return true, if the given {@link ITreeNode} object is a node otherwise false
@@ -343,7 +377,7 @@ public class ITreeNodeHandlerExtensions
 	 * @param <V>
 	 *            the generic type of the value
 	 * @param <T>
-	 *            the generic type of the id of the node
+	 *            the generic type of the concrete tree node
 	 * @param treeNode
 	 *            the tree node
 	 */
@@ -358,7 +392,7 @@ public class ITreeNodeHandlerExtensions
 	 * @param <V>
 	 *            the generic type of the value
 	 * @param <T>
-	 *            the generic type of the id of the node
+	 *            the generic type of the concrete tree node
 	 * @param treeNode
 	 *            the tree node
 	 */
@@ -375,7 +409,7 @@ public class ITreeNodeHandlerExtensions
 	 * @param <V>
 	 *            the generic type of the value
 	 * @param <T>
-	 *            the generic type of the id of the node
+	 *            the generic type of the concrete tree node
 	 * @param parentTreeNode
 	 *            the parent tree node
 	 * @param child
@@ -397,7 +431,7 @@ public class ITreeNodeHandlerExtensions
 	 * @param <V>
 	 *            the generic type of the value
 	 * @param <T>
-	 *            the generic type of the id of the node
+	 *            the generic type of the concrete tree node
 	 * @param parentTreeNode
 	 *            the parent tree node
 	 * @param children
@@ -418,7 +452,7 @@ public class ITreeNodeHandlerExtensions
 	 * @param <V>
 	 *            the generic type of the value
 	 * @param <T>
-	 *            the generic type of the id of the node
+	 *            the generic type of the concrete tree node
 	 * @param treeNode
 	 *            the tree node
 	 * @return the child count
@@ -434,7 +468,7 @@ public class ITreeNodeHandlerExtensions
 	 * @param <V>
 	 *            the generic type of the value
 	 * @param <T>
-	 *            the generic type of the id of the node
+	 *            the generic type of the concrete tree node
 	 * @param treeNode
 	 *            the tree node
 	 * @return true, if the given {@link ITreeNode} object has children otherwise false
@@ -451,7 +485,7 @@ public class ITreeNodeHandlerExtensions
 	 * @param <V>
 	 *            the generic type of the value
 	 * @param <T>
-	 *            the generic type of the id of the node
+	 *            the generic type of the concrete tree node
 	 * @param treeNode
 	 *            the tree node
 	 *
@@ -466,62 +500,13 @@ public class ITreeNodeHandlerExtensions
 	}
 
 	/**
-	 * Accepts the given visitor that provides a custom algorithm for processing all elements
-	 *
-	 * @param <V>
-	 *            the generic type of the value
-	 * @param <T>
-	 *            the generic type of the id of the node
-	 * @param treeNode
-	 *            the tree node
-	 * @param visitor
-	 *            the visitor
-	 */
-	public static <V, T extends ITreeNode<V, T>> void accept(final @NonNull T treeNode,
-		final @NonNull Visitor<T> visitor)
-	{
-		ITreeNodeHandlerExtensions.accept(treeNode, visitor, false);
-	}
-
-	/**
-	 * Accepts the given visitor that provides a custom algorithm for processing all elements
-	 *
-	 * @param <V>
-	 *            the generic type of the value
-	 * @param <T>
-	 *            the generic type of the id of the node
-	 * @param treeNode
-	 *            the tree node
-	 * @param visitor
-	 *            the visitor
-	 * @param visitBefore
-	 *            the flag if this flag is true the visit of the given {@link ITreeNode} object is
-	 *            before visit the children otherwise the visit is after visit the children
-	 */
-	public static <V, T extends ITreeNode<V, T>> void accept(final @NonNull T treeNode,
-		final @NonNull Visitor<T> visitor, final boolean visitBefore)
-	{
-		boolean visitAfter = !visitBefore;
-		if (visitBefore)
-		{
-			visitor.visit(treeNode);
-		}
-		treeNode.getChildren()
-			.forEach(child -> ITreeNodeHandlerExtensions.accept(child, visitor, visitBefore));
-		if (visitAfter)
-		{
-			visitor.visit(treeNode);
-		}
-	}
-
-	/**
 	 * Find all {@link ITreeNode} objects from the first given {@link ITreeNode} object that serves
 	 * as the search target, that have the same value as the given value
 	 *
 	 * @param <V>
 	 *            the generic type of the value
 	 * @param <T>
-	 *            the generic type of the id of the node
+	 *            the generic type of the concrete tree node
 	 * @param treeNode
 	 *            the tree node
 	 * @param value
@@ -561,7 +546,7 @@ public class ITreeNodeHandlerExtensions
 	 * @param <V>
 	 *            the generic type of the value
 	 * @param <T>
-	 *            the generic type of the id of the node
+	 *            the generic type of the concrete tree node
 	 * @param treeNode
 	 *            the tree node
 	 * @param value
@@ -585,7 +570,7 @@ public class ITreeNodeHandlerExtensions
 	 * @param <V>
 	 *            the generic type of the value
 	 * @param <T>
-	 *            the generic type of the id of the node
+	 *            the generic type of the concrete tree node
 	 * @param treeNode
 	 *            the tree node
 	 * @param descendantCandidate
@@ -610,7 +595,7 @@ public class ITreeNodeHandlerExtensions
 	 * @param <V>
 	 *            the generic type of the value
 	 * @param <T>
-	 *            the generic type of the id of the node
+	 *            the generic type of the concrete tree node
 	 * @param treeNode
 	 *            the tree node
 	 * @param treeNodes
@@ -631,7 +616,7 @@ public class ITreeNodeHandlerExtensions
 	 * @param <V>
 	 *            the generic type of the value
 	 * @param <T>
-	 *            the generic type of the id of the node
+	 *            the generic type of the concrete tree node
 	 * @param treeNode
 	 *            the tree node
 	 *