+++ src/test/java/io/github/astrapi69/gen/tree/element/TreeElement.java
@@ -24,11 +24,17 @@
  */
 package io.github.astrapi69.gen.tree.element;
 
-import lombok.*;
+import java.io.Serializable;
+
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.EqualsAndHashCode;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+import lombok.ToString;
 import lombok.experimental.FieldDefaults;
 import lombok.experimental.SuperBuilder;
-
-import java.io.Serializable;
 
 /**
  * The class {@link TreeElement} represents as the name already presume a tree element