+++ src/test/java/io/github/astrapi69/gen/tree/element/TreeElementNodeTest.java
@@ -24,14 +24,15 @@
  */
 package io.github.astrapi69.gen.tree.element;
 
-import io.github.astrapi69.gen.tree.TreeNode;
+import static org.testng.AssertJUnit.assertNotNull;
+
 import org.meanbean.lang.Factory;
 import org.meanbean.test.BeanTester;
 import org.meanbean.test.Configuration;
 import org.meanbean.test.ConfigurationBuilder;
 import org.testng.annotations.Test;
 
-import static org.testng.AssertJUnit.assertNotNull;
+import io.github.astrapi69.gen.tree.TreeNode;
 
 /**
  * The unit test class for the class {@link TreeElementNode}
@@ -78,7 +79,6 @@ public class TreeElementNodeTest
 
 			}).build();
 		final BeanTester beanTester = new BeanTester();
-		beanTester.addCustomConfiguration(TreeNode.class, configuration);
-		beanTester.testBean(TreeNode.class);
+		beanTester.testBean(TreeNode.class, configuration);
 	}
 }